requirements.txt
pandas
numpy
scikit-learn
streamlit
joblib
openai
pdfplumber
pytesseract
python-docx
Pillow
psycopg2-binary
SQLAlchemy
python-dateutil
chromadb
sqlalchemy
langchain
langchain-ollama
langchain-chroma
ollama